backend/requirements.txt
accelerate==1.10.1
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
appdirs==1.4.4
attrs==25.3.0
authlib==1.6.3
azure-ai-documentintelligence==1.0.2
azure-core==1.35.0
azure-identity==1.24.0
beautifulsoup4==4.13.5
cachetools==5.5.2
certifi==2025.8.3
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.3
click==8.2.1
cobble==0.1.4
colorama==0.4.6
coloredlogs==15.0.1
contourpy
cryptography==45.0.7
cycler==0.12.1
cyclopts==3.24.0
dataclasses-json==0.6.7
defusedxml==0.7.1
deprecation==2.1.0
distro==1.9.0
dnspython==2.8.0
docstring-parser==0.17.0
docutils==0.22
email-validator==2.3.0
et-xmlfile==2.0.0
exceptiongroup==1.3.0
fastapi==0.116.1
fastmcp==2.12.2
filelock==3.19.1
filetype==1.2.0
flatbuffers==25.2.10
fonttools==4.59.2
frozenlist==1.7.0
fsspec==2025.9.0
gitdb==4.0.12
gitpython==3.1.45
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.181.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
greenlet==3.2.4
groq==0.31.1
grpcio==1.74.0
grpcio-status==1.71.2
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.30.0
httpx==0.28.1
httpx-sse==0.4.1
huggingface-hub==0.35.0
humanfriendly==10.0
hyperframe==6.1.0
idna==3.10
importlib-metadata==8.7.0
isodate==0.7.2
jinja2==3.1.6
jiter==0.10.0
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-path==0.3.4
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
langchain==0.3.27
langchain-community==0.3.29
langchain-core==0.3.75
langchain-experimental==0.3.4
langchain-google-genai==2.0.10
langchain-groq==0.3.7
langchain-markitdown==0.1.8
langchain-openai==0.3.32
langchain-pymupdf4llm==0.4.1
langchain-text-splitters==0.3.11
langsmith==0.4.26
lazy-object-proxy==1.12.0
lxml==6.0.1
magika==0.6.2
mammoth==1.10.0
markdown-it-py==4.0.0
markdownify==1.2.0
markitdown==0.1.3
markupsafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.6
mcp==1.13.1
mdurl==0.1.2
mermaid-py==0.8.0
more-itertools==10.8.0
mpmath==1.3.0
msal==1.33.0
msal-extensions==1.3.1
multidict==6.6.4
mypy-extensions==1.1.0
neo4j==5.28.2
networkx
numpy==2.2.6
olefile==0.47
onnxruntime==1.20.1
openai==1.107.1
openapi-core==0.19.5
openapi-pydantic==0.5.1
openapi-schema-validator==0.6.3
openapi-spec-validator==0.7.2
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
openpyxl==3.1.5
orjson==3.11.3
packaging==25.0
pandas==2.3.2
parse==1.20.2
pathable==0.4.4
pdfminer-six==20250506
pgvector==0.4.1
pillow==11.3.0
postgrest==1.1.1
pptx2img==0.1.0
propcache==0.3.2
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1-modules==0.4.2
pycparser==2.22
pydantic==2.11.7
pydantic-core==2.33.2
pydantic-settings==2.10.1
pydub==0.25.1
pyee==11.1.1
pygments==2.19.2
pyjwt==2.10.1
pymupdf==1.26.4
pymupdf4llm==0.0.27
pyparsing==3.2.3
pypdfium2==4.30.0
pyperclip==1.9.0
pyautogui==0.9.54
pymupdf>=1.24.10
pyreadline3==3.5.4
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-dotenv==1.1.1
python-multipart==0.0.20
python-pptx==1.0.2
pytz==2025.2
pywin32==311
pyyaml==6.0.2
rank-bm25==0.2.2
realtime==2.7.0
referencing==0.36.2
regex==2025.9.1
requests==2.32.5
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rich==14.1.0
rich-rst==1.3.1
rpds-py==0.27.1
rsa==4.9.1
safetensors==0.6.2
scikit-learn==1.7.2
scipy
sentence-transformers==5.1.1
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.8
speechrecognition==3.14.3
sqlalchemy==2.0.43
sse-starlette==3.0.2
starlette==0.47.3
storage3==0.12.1
strenum==0.4.15
structlog==25.4.0
supabase==2.18.1
supabase-auth==2.12.3
supabase-functions==0.10.1
sympy==1.14.0
tabula-py==2.10.0
tabulate==0.9.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.11.0
tokenizers==0.22.1
tqdm==4.67.1
transformers==4.56.2
typer==0.17.4
typing-extensions==4.15.0
typing-inspect==0.9.0
typing-inspection==0.4.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.35.0
websockets==15.0.1
werkzeug==3.1.1
wheel==0.45.1
xlrd==2.0.2
xlsxwriter==3.2.5
yarl==1.20.1
youtube-transcript-api==1.0.3
zipp==3.23.0
zstandard==0.24.0